Move configuration construction into the resources package.
